go: don't use temp files for compiled tests (#2184)

+++ cli/cmd/encore/test.go
@@ -96,15 +96,20 @@ func runTests(appRoot, testDir string, args []string, traceFile string, codegenD
 		converter = convertTestEventOutputOnly(converter)
 	}
 
-	daemon := setupDaemon(ctx)
-
-	tempDir, err := os.MkdirTemp("", "encore-test")
-	if err != nil {
-		return 1, fmt.Errorf("couldn't create temp dir for test: %w", err)
+	var tempDir string
+	// only use temp dir if we are not compiling tests
+	if !slices.Contains(args, "-o") && !slices.Contains(args, "-c") {
+		var err error
+		tempDir, err = os.MkdirTemp("", "encore-test")
+		if err != nil {
+			return 1, fmt.Errorf("couldn't create temp dir for test: %w", err)
+		}
+		defer func() {
+			_ = os.RemoveAll(tempDir)
+		}()
 	}
-	defer func() {
-		_ = os.RemoveAll(tempDir)
-	}()
+
+	daemon := setupDaemon(ctx)
 
 	// Is this a node package?
 	packageJsonPath := filepath.Join(appRoot, "package.json")
